+++ zakura-network/src/peer/load_tracked_client.rs
@@ -2,7 +2,7 @@
 //! reported protocol version.
 
 use std::{
-    net::{IpAddr, SocketAddr},
+    net::IpAddr,
     sync::Arc,
     task::{Context, Poll},
 };
@@ -14,8 +14,8 @@ use tower::{
 
 use crate::{
     constants::{EWMA_DECAY_TIME_NANOS, EWMA_DEFAULT_RTT},
-    peer::{Client, ConnectedAddr, ConnectionInfo},
-    protocol::external::{canonical_socket_addr, types::Version},
+    peer::{Client, ConnectionInfo},
+    protocol::external::types::Version,
 };
 
 /// A client service wrapper that keeps track of its load.
@@ -57,13 +57,9 @@ impl LoadTrackedClient {
 
     /// Returns true if this peer connected directly to us from `ip`.
     pub fn is_inbound_direct_from_ip(&self, ip: &IpAddr) -> bool {
-        let expected_ip = canonical_socket_addr(SocketAddr::new(*ip, 0)).ip();
-
-        matches!(
-            self.connection_info.connected_addr,
-            ConnectedAddr::InboundDirect { addr }
-                if canonical_socket_addr(addr.remove_socket_addr_privacy()).ip() == expected_ip
-        )
+        self.connection_info
+            .connected_addr
+            .is_inbound_direct_from_ip(ip)
     }
 }